SRP cleanup
