+++ src/AppBundle/Twig/PremiumContentExtension.php
@@ -5,42 +5,38 @@
  */
 namespace AppBundle\Twig;
 
-use DOMDocument;
-use DOMNode;
-use SebastianBergmann\CodeCoverage\Report\PHP;
+use AppBundle\PremiumContent\HtmlRenderer;
+use AppBundle\PremiumContent\PermissionChecker;
 use Twig_Extension;
 use Twig_SimpleFilter;
 use Twig_SimpleFunction;
-use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
-use eZ\Publish\API\Repository\Repository as RepositoryInterface;
-use eZ\Publish\API\Repository\Values\User\User;
 
 /**
  * Twig helper for premium content.
  */
 class PremiumContentExtension extends Twig_Extension
 {
-    /** @var \eZ\Publish\API\Repository\Repository */
-    private $repository;
+    /** @var \AppBundle\PremiumContent\HtmlRenderer */
+    private $htmlRenderer;
 
-    /** @var \Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface */
-    private $tokenStorage;
+    /** @var \AppBundle\PremiumContent\PermissionChecker */
+    private $permissionChecker;
 
-    /** @var array */
+    /** @var int[] */
     private $allowedUserGroupsLocationIds;
 
     /**
-     * @param \eZ\Publish\API\Repository\Repository $repository
-     * @param \Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface $tokenStorage
+     * @param \AppBundle\PremiumContent\HtmlRenderer $htmlRenderer
+     * @param \AppBundle\PremiumContent\PermissionChecker $permissionChecker
      * @param array $allowedUserGroupsLocationIds
      */
     public function __construct(
-        RepositoryInterface $repository,
-        TokenStorageInterface $tokenStorage,
+        HtmlRenderer $htmlRenderer,
+        PermissionChecker $permissionChecker,
         array $allowedUserGroupsLocationIds
     ) {
-        $this->repository = $repository;
-        $this->tokenStorage = $tokenStorage;
+        $this->htmlRenderer = $htmlRenderer;
+        $this->permissionChecker = $permissionChecker;
         $this->allowedUserGroupsLocationIds = $allowedUserGroupsLocationIds;
     }
 
@@ -88,38 +84,7 @@ class PremiumContentExtension extends Twig_Extension
      */
     public function previewPremiumContent($document, $numberOfDisplayedElements = 2): string
     {
-        $doc = new DOMDocument();
-        libxml_use_internal_errors(true);
-        $doc->loadHTML($document, LIBXML_HTML_NOIMPLIED | LIBXML_HTML_NODEFDTD);
-
-        $childNodes = $doc->documentElement->childNodes;
-        $numberOfNodesToKeep = 0;
-        $importantNodes = 0;
-
-        /** @var \DOMElement $node */
-        foreach ($childNodes as $node) {
-            ++$numberOfNodesToKeep;
-
-            if ($node->nodeType === XML_ELEMENT_NODE) {
-                ++$importantNodes;
-            }
-
-            if ($importantNodes >= $numberOfDisplayedElements) {
-                break;
-            }
-        }
-
-        while ($childNodes->length > $numberOfNodesToKeep) {
-            $lastNode = $childNodes->item($childNodes->length - 1);
-            $lastNode->parentNode->removeChild($lastNode);
-        }
-
-        return $this->removeXmlHeader($doc->saveXML());
-    }
-
-    private function removeXmlHeader(string $html)
-    {
-        return str_replace('<?xml version="1.0" standalone="yes"?>' . "\n", null, $html);
+        return $this->htmlRenderer->renderElements($document, $numberOfDisplayedElements);
     }
 
     /**
@@ -135,42 +100,6 @@ class PremiumContentExtension extends Twig_Extension
             return $hasAccess;
         }
 
-        if (false === $token = $this->tokenStorage->getToken()) {
-            return false;
-        }
-
-        if (!\is_object($token->getUser())) {
-            return false;
-        }
-
-        $userGroups = $this->loadUserGroups($token->getUser()->getAPIUser());
-
-        foreach ($userGroups as $userGroup) {
-            if (\in_array(
-                $userGroup->contentInfo->mainLocationId,
-                $this->allowedUserGroupsLocationIds,
-                true
-            )) {
-                return $hasAccess = true;
-            }
-        }
-
-        return $hasAccess = false;
-    }
-
-    /**
-     * Loads User Groups of User, regardless to user limitations.
-     *
-     * @param \eZ\Publish\API\Repository\Values\User\User $apiUser
-     *
-     * @return \eZ\Publish\API\Repository\Values\User\UserGroup[]
-     */
-    private function loadUserGroups(User $apiUser): array
-    {
-        return $this->repository->sudo(
-            function (RepositoryInterface $repository) use ($apiUser) {
-                return $repository->getUserService()->loadUserGroupsOfUser($apiUser);
-            }
-        );
+        return $hasAccess = $this->permissionChecker->isCurrentUserInOneOfTheGroups($this->allowedUserGroupsLocationIds);
     }
 }
